requirements.txt
streamlit
langchain-core
langchain-community
langchain-text-splitters
langchain-groq
faiss-cpu
pdfplumber
python-dotenv
